Add verification of Merkle proofs to checkings
